requirements.txt
# Core ML Libraries
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
pillow>=10.0.0
scikit-learn>=1.3.0
pandas>=2.0.0

# Data Processing
opencv-python>=4.8.0
albumentations>=1.3.0

# Experiment Tracking
mlflow>=2.8.0

# Data Versioning
dvc>=3.30.0
dvc-s3>=3.0.0  # For S3 remote storage

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0

# Monitoring & Logging
prometheus-client>=0.19.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Utilities
python-dotenv>=1.0.0
PyYAML>=6.0.0
click>=8.1.0
tqdm>=4.66.0
requests>=2.31.0

# Development
black>=23.11.0
isort>=5.12.0
flake8>=6.1.0
mypy>=1.7.0
